chore: Dimensions calculation move to provider
