+++ projects/Mallard/src/screens/article/slider/index.tsx
@@ -9,7 +9,7 @@ import { getColor } from 'src/helpers/transform'
 import { getAppearancePillar } from 'src/hooks/use-article'
 import { useDismissArticle } from 'src/hooks/use-dismiss-article'
 import { useSetNavPosition } from 'src/hooks/use-nav-position'
-import { useDimensions } from 'src/hooks/use-screen'
+import { useDimensions } from 'src/hooks/use-config-provider'
 import { ArticleNavigationProps } from 'src/navigation/helpers/base'
 import { ArticleSpec, getArticleDataFromNavigator } from '../../article-screen'
 import { ArticleScreenBody, OnIsAtTopChange } from '../body'
